[ws] Simplify use of per-ws roots

+++ widgets/lib/widget-desktop.nu
@@ -17,22 +17,26 @@ def render [] {
   print --no-newline (tput cup 0 0)
 
   let ws = (wm ws)
-  let root = ($ws | ws root)
+  let root = ($ws | ws meta | get --ignore-errors root)
+  let wsid = $" ('󰕮 ' | fade) ($ws | ws meta | get name) "
 
-  cd $root
+  if $root != null {
+    cd $root
 
-  let wsid = $" ('󰕮 ' | fade) ($ws | get name) "
-  let wsnlen = 1 + 2 + 1 + ($ws | get name | into string | str length) + 1 # space, icon, space, id len, space
+    let wsnlen = 1 + 2 + 1 + ($ws | ws meta | get name | str length) + 1 # space, icon, space, id len, space
+    let cols = (term size | get columns)
+    let wsrlen = $cols - $wsnlen
 
-  let cols = (term size | get columns)
-  let wsrlen = $cols - $wsnlen
-  let wsroot = (
-    $" (' ' | fade) ($root | str replace $"($env.HOME)/" "")"
-    | strx truncate $wsrlen
-    | fill --width $wsrlen
-  )
+    let wsroot = (
+      $" (' ' | fade) ($root | str replace $"($env.HOME)/" "")"
+      | strx truncate $wsrlen
+      | fill --width $wsrlen
+    )
 
-  print $"($wsroot)($wsid)"
+    print $"($wsroot)($wsid)"
+  } else {
+    print $"($wsid)(ansi erase_line_from_cursor_to_end)"
+  }
 
   let branch = (git branch --show-current e> /dev/null | str trim)
